Support mobile access to playground site (#73)
## Motivation for the change, related issues

This PR updates the private web app to fall back to wpcom OAuth when a
request is not A8C-proxied. This way, users connecting via mobile
devices can still access the Playground instance.

Closes #53

## Implementation details

We add a new script `site-privacy.php` which helps enforce privacy by
doing proxy and oauth checks. This enforcement only works for requests
that are served via PHP, but that is sufficient for the private
Playground instance. We are serving the app for internal purposes, but
there is nothing secret about the app.

## Testing Instructions (or ideally a Blueprint)

- Temporarily disable the branch and user protections in the deployment
workflow
- Run it to deploy the site
- Confirm access is allowed for:
  - Users connecting via A8C proxy
  - Automatticians signing in via OAuth 
- Confirm Access is denied for a non-Automattic test user who attempts
to sign in via OAuth