[PHP] Improved support for spawning subprocesses (#2259)
This PR improves code surrounding spawning PHP subprocesses to:

* fsync() the TTY output buffer as it receives new chunks
* Use process.output.write for TTY messages instead of console.log()
* Ensure the child process emits the `spawn` event to tell the parent
process it's not hanging. It's the same `spawn` event as Node.js emits
when you use `child_process`.
* Prevents duplicate resolve() or reject() calls when spawning the
process
* Add a spawn timeout to avoid blocking the process with infinite
waiting
* When the process exits prematurely, emits it's pid if it got one

## Testing instructions

Confirm the CI checks passed. I did not find a good way of testing
everything, but we're at least testing for the process not hanging
indefinitely.